Issue #29 by default the Docker run will make use of in-memory mounts for the application and data path in the standard Postgres Docker image. If in-memory must be disabled, it can be done in the properties. Custom paths can be passed as a list under the inMemoryMountDestinations list.
The init process has been revamped to make it less reliable on the first lazy call of getProperties().

The in-memory solution requires a dynamic number of mount mappings. The solution is to insert a template (IN_MEMORY_TEMPLATE), which is transformed into the required number of --mount mappings for Docker. IN_MEMORY_TEMPLATE is hardcoded for now. Not sure if it is sensible to make this customizable.